Pastry knows how to add nodes to its routing table
In this commit pastry is updated to be able to augment
its routing table with new nodes. They are added where
apropriate. When receiving routing tables from peers
the pastry node checks the distance to the nodes in the
routing table and updates its own routing table accordingly
if any of the new nodes are better than the ones it already
knows about.

fs.app has additionally been updated to include pastry_app,
the main pastry application program (not yet implemented).