HubApp updates
Hub app now acts as a rendevouz point through TCP instead over
HTTP on top of TCP. This simplifies the protocol significatly.

Additionally it checks if nodes are alive and removes dead
nodes from its state table so it doesn't have to be reset
manually anymore.

Also started developing a hub web interface. It currently
only lists the nodes that are running.