Pastry and Chord transition to using pid
Now the chord and pastry servers reference their
own pids when calling themselves, and several
nodes can be run in the same vm.

Additionally the controller server has been
updated to start and stop nodes through a
small and simple state machine.

Friendsearch is also updated to reference
a current Dht when the Dht type is swapped.