Hub_app node temporarily updates controller state
When a new node joins the node port is added to the list
of ports a controller maintains temporarily. This is needed
in the case where there are few nodes in a network. So few
indeed that there wouldn't be enough to let them properly
rendevouz. The controller state will be updated again
by the next ping/pong session to ensure that the state
is correct