Additions to logger and bug fix
The logger now logs the data consumption of the Dhts.
It distinguishes between data that is part of looking
up a key, and general traffic that is part of maintaining
the state of the Dht network.

Fixed a bug where the controller would crash if it received
a request to shut down a node when there were none to shut down.