Updated logparser and fixed logging bug in pastry
The log parser now takes into account the new log format
and produces CSVs that can be used to visualise the data.

A bug was fixed where pastry wouldn't log the data used
as part of lookups.

The hub now requires 5% of the nodes to stop the experiment
before it is ended.

Controllers now only say stop when it has a history of at
least 100 lookups.