Bug fix in Chord routing
I finally found a major bug in the closest_preceding_finger
function, where the ranges were not correctly checked.
This bug has caused all sort of mayhem, and resultet
in the correct nodes never being found, or the chord
network ending up with looking for a match indefinitely.

Additionally the routing code was cleaned up significantly,
and now much more closely mirrors the pseudo code in the chord paper