Improved experimental runner
The code base was significantly trimmed.
All the previous experimental options were removed
and now the only one remaining is the one allowing
me to run a fixed rate experiment for a given period
of time.
The ability for nodes to ask the command center (hub)
to stop, has also been removed, and changes have also
been made so the rate of requests should now be more
closely followed!