Reactivate the pastry locality heuristic
The Pastry locality heuristic had been temporarily
deactivated to allow me to test the actual performance
of the heuristic.

Test results:
Not sufficient test data to make conclusive remarks,
but it seems like large number of nodes on the same
machine make the heuristic perform really badly,
which is a case in which you would expect the heuristic
to perform extremely well!

I guess the heuristic itself might be bad, and that
it should be replaced with a different one!